+++ src/Pages/Seokmoon/Main/Main.js
@@ -1,16 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import { FOOTER_LIST } from './uiData.js';
-import UserProfile from './UserProfile.js';
-import Comments from './Comments.js';
-// import MainFeed from './MainFeed.js';
+import UserProfile from '../Main/components/UserProfile';
 import Nav from '../../../components/Nav/Nav';
+import Feed from '../Main/components/Feed';
 import './Main.scss';
 
 function Main() {
-  const [userName] = useState('tjrans9248');
-  const [feedComments, setFeedComments] = useState([]);
-  const [comment, setComment] = useState('');
-
   const [feedList, setFeedList] = useState([]);
 
   useEffect(() => {
@@ -18,17 +13,6 @@ function Main() {
       .then(response => response.json())
       .then(result => setFeedList(result));
   }, []);
-
-  const post = e => {
-    const copyFeedComments = [...feedComments];
-    copyFeedComments.push(comment);
-    setFeedComments(copyFeedComments);
-    setComment('');
-  };
-
-  const onChange = e => {
-    setComment(e.target.value);
-  };
 
   return (
     <div className="background">
@@ -39,97 +23,14 @@ function Main() {
         <div className="main-container">
           <div className="main-feeds">
             <div className="article-box">
-              {feedList.map(feedInfo => {
-                return (
-                  <div key={feedInfo.id}>
-                    <div className="feed-nav">
-                      <div>
-                        <div className="main-image-box">
-                          <img
-                            src={feedInfo.profileImg}
-                            className="main-image"
-                            alt=""
-                          />
-                        </div>
-                      </div>
-                      <div>
-                        <span className="user-id">{feedInfo.userId}</span>
-                        <i className="fa-solid fa-ellipsis i-dot" />
-                      </div>
-                    </div>
-
-                    <div className="feed-main">
-                      <img
-                        src={feedInfo.imageUrl}
-                        alt=""
-                        className="feed-image"
-                      />
-                    </div>
-
-                    <div className="feed-footer-back">
-                      <div className="feed-footer">
-                        <div>
-                          <i className="fa-solid fa-heart feed-heart feed-icon" />
-                          <i className="fa-solid fa-comment feed-icon" />
-                          <i className="fa-solid fa-arrow-up-from-bracket feed-icon" />
-                        </div>
-
-                        <div>
-                          <i className="fa-solid fa-bookmark feed-icon" />
-                        </div>
-                      </div>
-
-                      <div>
-                        <strong>{feedInfo.userId}</strong>님{' '}
-                        <strong>외 10명</strong>이 좋아합니다.
-                      </div>
-
-                      <div>
-                        <p className="comment">
-                          <strong>seokmoonYang</strong> 올드카가 굉장히
-                          멋있습니다....
-                        </p>
-                        <p className="comment">
-                          <strong>neceosecius</strong> 이건 얼마정도할까요???
-                        </p>
-                      </div>
-
-                      <div className="comment-input-box">
-                        {feedComments.map((commentArr, user) => {
-                          return (
-                            <div className="mainComments" key={user}>
-                              <Comments
-                                userName={userName}
-                                commentArr={commentArr}
-                              />
-                            </div>
-                          );
-                        })}
-                      </div>
-
-                      <div>
-                        <div className="border-top">
-                          <input
-                            className="input-comment"
-                            type="text"
-                            placeholder="댓글 달기..."
-                            style={{ fontSize: '18px' }}
-                            onChange={onChange}
-                            value={comment}
-                          />
-
-                          <input
-                            className="input-button"
-                            type="button"
-                            value="게시"
-                            onClick={post}
-                          />
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                );
-              })}
+              {feedList.map(feedInfo => (
+                <Feed
+                  profileImg={feedInfo.profileImg}
+                  imageUrl={feedInfo.imageUrl}
+                  userId={feedInfo.userId}
+                  key={feedInfo.id}
+                />
+              ))}
             </div>
             <div className="main-right">
               <div>
